move mail.js to folder Mail
